refactor: 디렉토리 표기법을 스네이크에서 카멜로 수정 수정
