+++ src/main/java/roomescape/reservation/domain/Reservation.java
@@ -3,7 +3,7 @@ package roomescape.reservation.domain;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import roomescape.globalException.BadRequestException;
-import roomescape.reservation_time.domain.ReservationTime;
+import roomescape.reservationTime.domain.ReservationTime;
 import roomescape.theme.domain.Theme;
 
 public class Reservation {
