+++ src/main/java/roomescape/reservation/domain/dto/ReservationResDto.java
@@ -1,7 +1,7 @@
 package roomescape.reservation.domain.dto;
 
 import java.time.LocalDate;
-import roomescape.reservation_time.domain.dto.ReservationTimeResDto;
+import roomescape.reservationTime.domain.dto.ReservationTimeResDto;
 import roomescape.theme.domain.dto.ThemeResDto;
 
 public record ReservationResDto(
