+++ src/main/java/roomescape/reservation/repository/ReservationRepositoryImpl.java
@@ -11,7 +11,7 @@ import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import roomescape.globalException.NotFoundException;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation_time.domain.ReservationTime;
+import roomescape.reservationTime.domain.ReservationTime;
 import roomescape.theme.domain.Theme;
 
 @Repository
