+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -8,10 +8,10 @@ import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.dto.ReservationReqDto;
 import roomescape.reservation.domain.dto.ReservationResDto;
 import roomescape.reservation.repository.ReservationRepository;
-import roomescape.reservation_time.ReservationTimeMapper;
-import roomescape.reservation_time.domain.ReservationTime;
-import roomescape.reservation_time.domain.dto.ReservationTimeResDto;
-import roomescape.reservation_time.repository.ReservationTimeRepository;
+import roomescape.reservationTime.ReservationTimeMapper;
+import roomescape.reservationTime.domain.ReservationTime;
+import roomescape.reservationTime.domain.dto.ReservationTimeResDto;
+import roomescape.reservationTime.repository.ReservationTimeRepository;
 import roomescape.theme.ThemeMapper;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.dto.ThemeResDto;
